test: 통합 테스트 medium 파일로 통합
